requirements.txt
#qbittorrent monitoring
qbittorrent-api

#plex monitoring
plexapi

#process monitor
psutil

#linux shutdown
dbus-python